feat: 유저선택 영상과 자동 생성 영상 기능구현

+++ backend/django_back/videos/utils.py
@@ -6,14 +6,14 @@ import boto3
 import os
 import random
 from stories.models import Story
+from videos.models import Video
 
-
-def make_video(user_id, video_number, image_urls, music_url):
+def make_video(capsule_id, video_number, image_urls, music_url):
     s3_client = boto3.client('s3')
     s3_resource = boto3.resource('s3')
     bucket_name = 'author-picture'
 
-    output_video_key = f'video-of-{user_id}-no{video_number}.mp4'
+    output_video_key = f'video-of-capsule{capsule_id}-no{video_number}.mp4'
 
     # Download and process images
     images = []
@@ -53,7 +53,7 @@ def make_video(user_id, video_number, image_urls, music_url):
         video_clip = video_clip.set_audio(audio_clip)
 
         # Generate final video with music
-        final_output = f'video_of_{user_id}_no{video_number}.mp4'
+        final_output = f'video-of-capsule{capsule_id}-no{video_number}.mp4'
         video_duration = len(images)  # Duration of the video in seconds
         video_clip = video_clip.subclip(0, video_duration)  # Truncate the video to match the duration of the images
         final_clip = mp.concatenate_videoclips([video_clip])
@@ -65,9 +65,12 @@ def make_video(user_id, video_number, image_urls, music_url):
         print('video uploaded to S3:', final_output)
 
         # Delete temporary files
-        os.remove(output_video_key)
-        os.remove('music.mp3')
-        os.remove(final_output)
+        if os.path.exists(output_video_key):
+            os.remove(output_video_key)
+        if os.path.exists('music.mp3'):
+            os.remove('music.mp3')
+        if os.path.exists(final_output):
+            os.remove(final_output)
 
         return f'https://{bucket_name}.s3.ap-northeast-2.amazonaws.com/{final_output}'
 
@@ -79,17 +82,11 @@ def make_video(user_id, video_number, image_urls, music_url):
         if os.path.exists(output_video_key):
             os.remove(output_video_key)
 
-
-# 오류 나면 final_output = f'final_video_of_{user_id}_no_{video_number}.mp4'
-# 이걸 video_of_{user_id}_no_{video_number}.mp4 이렇게 바꿔서임
-
-
 def random_video_url_maker(capsule, stories):
     story_id_list = []
     for story in stories:
         story_id_list.append(story.story_id)
-
-
+        print(story.story_title)
     story_id_list.sort()
     story_count = len(story_id_list)
 
@@ -97,25 +94,42 @@ def random_video_url_maker(capsule, stories):
     video_image_list_ready = []
 
     if story_count < capsule.limit_count:
-        for i in range(0, story_count):
+        for i in range(story_count):
             for j in range(2):
                 video_image_list_ready.append(story_id_list[i])
-        print("일열 : ", video_image_list_ready)
     else:
         while len(video_image_list_ready) < capsule.limit_count:
-            random_number = random.randint(0, len(story_id_list) - 1)
+            random_number = random.randint(1, len(story_id_list) - 1)
             if random_number in unique_values:
                 continue
             video_image_list_ready.append(random_number)
             unique_values.add(random_number)
-        print("랜덤 : ", video_image_list_ready)
 
     video_image_list_ready.sort()
     video_image_url_list = []
     for story_id in video_image_list_ready:
+        story = Story.objects.get(story_id=story_id)
         for j in range(2):
-            story = Story.objects.get(story_id=story_id)
             # 비디오 제작에 넘길 이미지 url 배열
             video_image_url_list.append(story.story_img_url)
 
     return video_image_url_list
+
+
+def default_video_maker(capsule, music):
+    stories = Story.objects.filter(capsule_id=capsule.capsule_id)
+
+    video_image_url_list_final = random_video_url_maker(capsule, stories)
+    music_url = music.music_url
+
+    # 캡슐 비디오 개수로 비디오 url 만듦 (비디오 url은 video_of_{user_id}_no{video_count})
+    video_count = Video.objects.filter(capsule=capsule.capsule_id).count() + 1
+
+    # s3 업로드 용 함수
+    return make_video(capsule.capsule_id, video_count, video_image_url_list_final, music_url)  # 회원 아이디, 회원 비디오 개수,
+
+
+def user_choice_video_url(capsule, music, user_choice_list):
+    video_count = Video.objects.filter(capsule=capsule.capsule_id).count() + 1
+    music_url = music.music_url
+    return make_video(capsule.capsule_id, video_count, user_choice_list, music_url)
